Move more functions to tools file
