+++ ApplicationLibCode/ProjectDataModel/RimEclipseResultDefinition.cpp
@@ -20,37 +20,28 @@
 
 #include "RimEclipseResultDefinition.h"
 
-#include "RiaColorTables.h"
-#include "RiaColorTools.h"
 #include "RiaLogging.h"
-#include "RiaNncDefines.h"
 #include "RiaQDateTimeTools.h"
 
 #include "RicfCommandObject.h"
 
 #include "RigActiveCellInfo.h"
-#include "RigAllanDiagramData.h"
 #include "RigCaseCellResultsData.h"
 #include "RigEclipseCaseData.h"
 #include "RigEclipseResultInfo.h"
 #include "RigFlowDiagResultAddress.h"
 #include "RigFlowDiagResults.h"
 #include "RigFormationNames.h"
-#include "RigVisibleCategoriesCalculator.h"
 
 #include "Rim3dView.h"
 #include "Rim3dWellLogCurve.h"
 #include "RimCellEdgeColors.h"
-#include "RimColorLegend.h"
-#include "RimColorLegendItem.h"
 #include "RimContourMapProjection.h"
 #include "RimEclipseCase.h"
 #include "RimEclipseCellColors.h"
 #include "RimEclipseContourMapProjection.h"
 #include "RimEclipseContourMapView.h"
 #include "RimEclipseFaultColors.h"
-#include "RimEclipseInputProperty.h"
-#include "RimEclipseInputPropertyCollection.h"
 #include "RimEclipsePropertyFilter.h"
 #include "RimEclipseResultCase.h"
 #include "RimEclipseResultDefinitionTools.h"
@@ -66,9 +57,6 @@
 #include "RimProject.h"
 #include "RimRegularLegendConfig.h"
 #include "RimReservoirCellResultsStorage.h"
-#include "RimSimWellInView.h"
-#include "RimSimWellInViewCollection.h"
-#include "RimTernaryLegendConfig.h"
 #include "RimViewLinker.h"
 #include "RimWellLogExtractionCurve.h"
 #include "RimWellLogTrack.h"
@@ -77,7 +65,6 @@
 #include "RimGridStatisticsPlot.h"
 #endif
 
-#include "cafCategoryMapper.h"
 #include "cafPdmFieldScriptingCapability.h"
 #include "cafPdmUiListEditor.h"
 #include "cafPdmUiToolButtonEditor.h"
@@ -1728,85 +1715,18 @@ QString RimEclipseResultDefinition::flowDiagResUiText( bool shortLabel, int maxT
 ///
 //--------------------------------------------------------------------------------------------------
 QList<caf::PdmOptionItemInfo> RimEclipseResultDefinition::calcOptionsForVariableUiFieldStandard( RiaDefines::ResultCatType resultCatType,
-                                                                                                 const RigCaseCellResultsData* results,
+                                                                                                 RigCaseCellResultsData* const results,
                                                                                                  bool showDerivedResultsFirst,
                                                                                                  bool addPerCellFaceOptionItems,
                                                                                                  bool ternaryEnabled )
 {
     CVF_ASSERT( resultCatType != RiaDefines::ResultCatType::FLOW_DIAGNOSTICS && resultCatType != RiaDefines::ResultCatType::INJECTION_FLOODING );
 
-    if ( results )
-    {
-        QList<caf::PdmOptionItemInfo> optionList;
-
-        QStringList cellCenterResultNames;
-        QStringList cellFaceResultNames;
-
-        for ( const QString& s : RimEclipseResultDefinitionTools::getResultNamesForResultType( resultCatType, results ) )
-        {
-            if ( s == RiaResultNames::completionTypeResultName() )
-            {
-                if ( results->timeStepDates().empty() ) continue;
-            }
-
-            if ( RiaResultNames::isPerCellFaceResult( s ) )
-            {
-                cellFaceResultNames.push_back( s );
-            }
-            else
-            {
-                cellCenterResultNames.push_back( s );
-            }
-        }
-
-        cellCenterResultNames.sort();
-        cellFaceResultNames.sort();
-
-        // Cell Center result names
-        for ( const QString& s : cellCenterResultNames )
-        {
-            optionList.push_back( caf::PdmOptionItemInfo( s, s ) );
-        }
-
-        if ( addPerCellFaceOptionItems )
-        {
-            for ( const QString& s : cellFaceResultNames )
-            {
-                if ( showDerivedResultsFirst )
-                {
-                    optionList.push_front( caf::PdmOptionItemInfo( s, s ) );
-                }
-                else
-                {
-                    optionList.push_back( caf::PdmOptionItemInfo( s, s ) );
-                }
-            }
-
-            // Ternary Result
-            if ( ternaryEnabled )
-            {
-                bool hasAtLeastOneTernaryComponent = false;
-                if ( cellCenterResultNames.contains( RiaResultNames::soil() ) )
-                    hasAtLeastOneTernaryComponent = true;
-                else if ( cellCenterResultNames.contains( RiaResultNames::sgas() ) )
-                    hasAtLeastOneTernaryComponent = true;
-                else if ( cellCenterResultNames.contains( RiaResultNames::swat() ) )
-                    hasAtLeastOneTernaryComponent = true;
-
-                if ( resultCatType == RiaDefines::ResultCatType::DYNAMIC_NATIVE && hasAtLeastOneTernaryComponent )
-                {
-                    optionList.push_front( caf::PdmOptionItemInfo( RiaResultNames::ternarySaturationResultName(),
-                                                                   RiaResultNames::ternarySaturationResultName() ) );
-                }
-            }
-        }
-
-        optionList.push_front( caf::PdmOptionItemInfo( RiaResultNames::undefinedResultName(), RiaResultNames::undefinedResultName() ) );
-
-        return optionList;
-    }
-
-    return QList<caf::PdmOptionItemInfo>();
+    return RimEclipseResultDefinitionTools::calcOptionsForVariableUiFieldStandard( resultCatType,
+                                                                                   results,
+                                                                                   showDerivedResultsFirst,
+                                                                                   addPerCellFaceOptionItems,
+                                                                                   ternaryEnabled );
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -1816,35 +1736,6 @@ void RimEclipseResultDefinition::setTernaryEnabled( bool enabled )
 {
     m_ternaryEnabled = enabled;
 }
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool operator<( const cvf::Color3ub first, const cvf::Color3ub second )
-{
-    if ( first.r() != second.r() ) return first.r() < second.r();
-    if ( first.g() != second.g() ) return first.g() < second.g();
-    if ( first.b() != second.b() ) return first.b() < second.b();
-
-    return false;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-class TupleCompare
-{
-public:
-    bool operator()( const std::tuple<QString, int, cvf::Color3ub>& t1, const std::tuple<QString, int, cvf::Color3ub>& t2 ) const
-    {
-        using namespace std;
-        if ( get<0>( t1 ) != get<0>( t2 ) ) return get<0>( t1 ) < get<0>( t2 );
-        if ( get<1>( t1 ) != get<1>( t2 ) ) return get<1>( t1 ) < get<1>( t2 );
-        if ( get<2>( t1 ) != get<2>( t2 ) ) return get<2>( t1 ) < get<2>( t2 );
-
-        return false;
-    }
-};
 
 //--------------------------------------------------------------------------------------------------
 ///
@@ -1860,380 +1751,20 @@ void RimEclipseResultDefinition::updateRangesForExplicitLegends( RimRegularLegen
     {
         if ( this->isFlowDiagOrInjectionFlooding() )
         {
-            CVF_ASSERT( currentTimeStep >= 0 );
-
-            double                   globalMin, globalMax;
-            double                   globalPosClosestToZero, globalNegClosestToZero;
-            RigFlowDiagResults*      flowResultsData = this->flowDiagSolution()->flowDiagResults();
-            RigFlowDiagResultAddress resAddr         = this->flowDiagResAddress();
-
-            flowResultsData->minMaxScalarValues( resAddr, currentTimeStep, &globalMin, &globalMax );
-            flowResultsData->posNegClosestToZero( resAddr, currentTimeStep, &globalPosClosestToZero, &globalNegClosestToZero );
-
-            double localMin, localMax;
-            double localPosClosestToZero, localNegClosestToZero;
-            if ( this->hasDynamicResult() )
+            if ( currentTimeStep >= 0 )
             {
-                flowResultsData->minMaxScalarValues( resAddr, currentTimeStep, &localMin, &localMax );
-                flowResultsData->posNegClosestToZero( resAddr, currentTimeStep, &localPosClosestToZero, &localNegClosestToZero );
-            }
-            else
-            {
-                localMin = globalMin;
-                localMax = globalMax;
-
-                localPosClosestToZero = globalPosClosestToZero;
-                localNegClosestToZero = globalNegClosestToZero;
-            }
-
-            CVF_ASSERT( legendConfigToUpdate );
-
-            legendConfigToUpdate->disableAllTimeStepsRange( true );
-            legendConfigToUpdate->setClosestToZeroValues( globalPosClosestToZero,
-                                                          globalNegClosestToZero,
-                                                          localPosClosestToZero,
-                                                          localNegClosestToZero );
-            legendConfigToUpdate->setAutomaticRanges( globalMin, globalMax, localMin, localMax );
-
-            if ( this->hasCategoryResult() )
-            {
-                RimEclipseView* eclView = nullptr;
-                this->firstAncestorOrThisOfType( eclView );
-
-                if ( eclView )
-                {
-                    std::set<std::tuple<QString, int, cvf::Color3ub>, TupleCompare> categories;
-
-                    std::vector<QString> tracerNames = this->flowDiagSolution()->tracerNames();
-                    int                  tracerIndex = 0;
-
-                    for ( const auto& tracerName : tracerNames )
-                    {
-                        cvf::Color3ub color( cvf::Color3::GRAY );
-
-                        RimSimWellInView* well =
-                            eclView->wellCollection()->findWell( RimFlowDiagSolution::removeCrossFlowEnding( tracerName ) );
-
-                        if ( well ) color = cvf::Color3ub( well->wellPipeColor() );
-
-                        categories.insert( std::make_tuple( tracerName, tracerIndex, color ) );
-                        ++tracerIndex;
-                    }
-
-                    std::vector<std::tuple<QString, int, cvf::Color3ub>> categoryVector;
-
-                    if ( m_showOnlyVisibleCategoriesInLegend )
-                    {
-                        std::set<int> visibleTracers =
-                            RigVisibleCategoriesCalculator::visibleFlowDiagCategories( *eclView, *flowResultsData, resAddr, currentTimeStep );
-                        for ( auto tupIt : categories )
-                        {
-                            int tracerIndex = std::get<1>( tupIt );
-                            if ( visibleTracers.count( tracerIndex ) ) categoryVector.push_back( tupIt );
-                        }
-                    }
-                    else
-                    {
-                        for ( auto tupIt : categories )
-                        {
-                            categoryVector.push_back( tupIt );
-                        }
-                    }
-
-                    legendConfigToUpdate->setCategoryItems( categoryVector );
-                }
+                RimEclipseResultDefinitionTools::updateLegendForFlowDiagnostics( this, legendConfigToUpdate, currentTimeStep );
             }
         }
         else
         {
-            if ( !rimEclipseCase ) return;
-
-            RigEclipseCaseData* eclipseCaseData = rimEclipseCase->eclipseCaseData();
-            if ( !eclipseCaseData ) return;
-
-            RigCaseCellResultsData* cellResultsData = eclipseCaseData->results( this->porosityModel() );
-            cellResultsData->ensureKnownResultLoaded( this->eclipseResultAddress() );
-
-            double globalMin, globalMax;
-            double globalPosClosestToZero, globalNegClosestToZero;
-
-            cellResultsData->minMaxCellScalarValues( this->eclipseResultAddress(), globalMin, globalMax );
-            cellResultsData->posNegClosestToZero( this->eclipseResultAddress(), globalPosClosestToZero, globalNegClosestToZero );
-
-            double localMin, localMax;
-            double localPosClosestToZero, localNegClosestToZero;
-            if ( this->hasDynamicResult() && currentTimeStep >= 0 )
-            {
-                cellResultsData->minMaxCellScalarValues( this->eclipseResultAddress(), currentTimeStep, localMin, localMax );
-                cellResultsData->posNegClosestToZero( this->eclipseResultAddress(), currentTimeStep, localPosClosestToZero, localNegClosestToZero );
-            }
-            else
-            {
-                localMin = globalMin;
-                localMax = globalMax;
-
-                localPosClosestToZero = globalPosClosestToZero;
-                localNegClosestToZero = globalNegClosestToZero;
-            }
-
-            CVF_ASSERT( legendConfigToUpdate );
-
-            legendConfigToUpdate->disableAllTimeStepsRange( false );
-            legendConfigToUpdate->setClosestToZeroValues( globalPosClosestToZero,
-                                                          globalNegClosestToZero,
-                                                          localPosClosestToZero,
-                                                          localNegClosestToZero );
-            legendConfigToUpdate->setAutomaticRanges( globalMin, globalMax, localMin, localMax );
-
-            if ( this->hasCategoryResult() )
-            {
-                if ( this->resultType() == RiaDefines::ResultCatType::ALLAN_DIAGRAMS )
-                {
-                    if ( this->resultVariable() == RiaResultNames::formationAllanResultName() )
-                    {
-                        const std::vector<QString> fnVector = eclipseCaseData->formationNames();
-                        std::vector<int>           fnameIdxes;
-                        for ( int i = 0; i < static_cast<int>( fnVector.size() ); i++ )
-                        {
-                            fnameIdxes.push_back( i );
-                        }
-
-                        cvf::Color3ubArray legendBaseColors = legendConfigToUpdate->colorLegend()->colorArray();
-
-                        cvf::ref<caf::CategoryMapper> formationColorMapper = new caf::CategoryMapper;
-                        formationColorMapper->setCategories( fnameIdxes );
-                        formationColorMapper->setInterpolateColors( legendBaseColors );
-
-                        const std::map<std::pair<int, int>, int>& formationCombToCategory =
-                            eclipseCaseData->allanDiagramData()->formationCombinationToCategory();
-
-                        std::vector<std::tuple<QString, int, cvf::Color3ub>> categories;
-                        for ( int frmNameIdx : fnameIdxes )
-                        {
-                            cvf::Color3ub formationColor = formationColorMapper->mapToColor( frmNameIdx );
-                            categories.emplace_back( std::make_tuple( fnVector[frmNameIdx], frmNameIdx, formationColor ) );
-                        }
-
-                        std::set<size_t> visibleAllanCategories;
-                        {
-                            RimEclipseView* eclView = nullptr;
-                            this->firstAncestorOrThisOfType( eclView );
-
-                            visibleAllanCategories = RigVisibleCategoriesCalculator::visibleAllanCategories( eclView );
-                        }
-
-                        for ( auto [formationPair, categoryIndex] : formationCombToCategory )
-                        {
-                            int frmIdx1 = formationPair.first;
-                            int frmIdx2 = formationPair.second;
-
-                            if ( visibleAllanCategories.count( categoryIndex ) == 0 ) continue;
-
-                            int fnVectorSize = static_cast<int>( fnVector.size() );
-                            if ( frmIdx1 >= fnVectorSize || frmIdx2 >= fnVectorSize ) continue;
-
-                            QString frmName1 = fnVector[frmIdx1];
-                            QString frmName2 = fnVector[frmIdx2];
-
-                            cvf::Color3f formationColor1 = cvf::Color3f( formationColorMapper->mapToColor( frmIdx1 ) );
-                            cvf::Color3f formationColor2 = cvf::Color3f( formationColorMapper->mapToColor( frmIdx2 ) );
-
-                            cvf::Color3ub blendColor = cvf::Color3ub( RiaColorTools::blendCvfColors( formationColor1, formationColor2 ) );
-
-                            categories.emplace_back( std::make_tuple( frmName1 + "-" + frmName2, categoryIndex, blendColor ) );
-                        }
-
-                        legendConfigToUpdate->setCategoryItems( categories );
-                    }
-                    else if ( this->resultVariable() == RiaResultNames::formationBinaryAllanResultName() )
-                    {
-                        std::vector<std::tuple<QString, int, cvf::Color3ub>> categories;
-                        categories.emplace_back( std::make_tuple( "Same formation", 0, cvf::Color3ub::BROWN ) );
-                        categories.emplace_back( std::make_tuple( "Different formation", 1, cvf::Color3ub::ORANGE ) );
-
-                        legendConfigToUpdate->setCategoryItems( categories );
-                    }
-                }
-                else if ( this->resultType() == RiaDefines::ResultCatType::DYNAMIC_NATIVE &&
-                          this->resultVariable() == RiaResultNames::completionTypeResultName() )
-                {
-                    const std::vector<int>& visibleCategories = cellResultsData->uniqueCellScalarValues( this->eclipseResultAddress() );
-
-                    std::vector<RiaDefines::WellPathComponentType> supportedCompletionTypes = { RiaDefines::WellPathComponentType::WELL_PATH,
-                                                                                                RiaDefines::WellPathComponentType::FISHBONES,
-                                                                                                RiaDefines::WellPathComponentType::PERFORATION_INTERVAL,
-                                                                                                RiaDefines::WellPathComponentType::FRACTURE };
-
-                    RiaColorTables::WellPathComponentColors colors = RiaColorTables::wellPathComponentColors();
-
-                    std::vector<std::tuple<QString, int, cvf::Color3ub>> categories;
-                    for ( auto completionType : supportedCompletionTypes )
-                    {
-                        if ( std::find( visibleCategories.begin(), visibleCategories.end(), static_cast<int>( completionType ) ) !=
-                             visibleCategories.end() )
-                        {
-                            QString categoryText = caf::AppEnum<RiaDefines::WellPathComponentType>::uiText( completionType );
-                            categories.push_back( std::make_tuple( categoryText, static_cast<int>( completionType ), colors[completionType] ) );
-                        }
-                    }
-
-                    legendConfigToUpdate->setCategoryItems( categories );
-                }
-                else
-                {
-                    auto uniqueValues = cellResultsData->uniqueCellScalarValues( this->eclipseResultAddress() );
-                    if ( this->eclipseResultAddress().resultCatType() == RiaDefines::ResultCatType::FORMATION_NAMES )
-                    {
-                        std::vector<QString> fnVector = eclipseCaseData->formationNames();
-                        uniqueValues.resize( fnVector.size() );
-                        std::iota( uniqueValues.begin(), uniqueValues.end(), 0 );
-                    }
-
-                    cvf::Color3ubArray legendBaseColors = legendConfigToUpdate->colorLegend()->colorArray();
-
-                    cvf::ref<caf::CategoryMapper> categoryMapper = new caf::CategoryMapper;
-                    categoryMapper->setCategories( uniqueValues );
-                    categoryMapper->setInterpolateColors( legendBaseColors );
-
-                    std::vector<int> visibleCategoryValues = uniqueValues;
-
-                    if ( m_showOnlyVisibleCategoriesInLegend )
-                    {
-                        RimEclipseView* eclView = nullptr;
-                        this->firstAncestorOrThisOfType( eclView );
-                        if ( eclView && eclView->showWindow() )
-                        {
-                            // Check if current result is cell result, and update the visible set of values
-                            // TODO: Can be extended to the separate geometry results (separate fault result, separate
-                            // intersection results), but this requires some refactoring
-                            if ( eclView->cellResult() == this )
-                            {
-                                std::set<int> visibleCategorySet = RigVisibleCategoriesCalculator::visibleCategories( eclView );
-
-                                visibleCategoryValues.clear();
-                                visibleCategoryValues.insert( visibleCategoryValues.begin(),
-                                                              visibleCategorySet.begin(),
-                                                              visibleCategorySet.end() );
-                            }
-                        }
-                    }
-                    std::vector<std::tuple<QString, int, cvf::Color3ub>> categoryVector;
-
-                    for ( auto value : visibleCategoryValues )
-                    {
-                        cvf::Color3ub categoryColor = categoryMapper->mapToColor( value );
-
-                        QString valueTxt;
-                        if ( this->resultType() == RiaDefines::ResultCatType::FORMATION_NAMES )
-                        {
-                            std::vector<QString> fnVector = eclipseCaseData->formationNames();
-
-                            if ( value < static_cast<int>( fnVector.size() ) )
-                            {
-                                valueTxt = fnVector[value];
-                            }
-                        }
-                        else
-                        {
-                            auto items = legendConfigToUpdate->colorLegend()->colorLegendItems();
-                            auto it    = std::find_if( items.begin(),
-                                                    items.end(),
-                                                    [value]( const RimColorLegendItem* const item )
-                                                    { return item->categoryValue() == value; } );
-                            if ( it != items.end() && !( *it )->categoryName().isEmpty() )
-                            {
-                                valueTxt = QString( "%1 (%2)" ).arg( ( *it )->categoryName() ).arg( value );
-                            }
-                            else
-                            {
-                                valueTxt = QString( "%1" ).arg( value );
-                            }
-                        }
-
-                        categoryVector.push_back( std::make_tuple( valueTxt, value, categoryColor ) );
-                    }
-                    legendConfigToUpdate->setCategoryItems( categoryVector );
-                }
-            }
+            RimEclipseResultDefinitionTools::updateCellResultLegend( this, legendConfigToUpdate, currentTimeStep );
         }
     }
 
-    // Ternary legend update
+    if ( isTernarySaturationSelected() )
     {
-        if ( !rimEclipseCase ) return;
-
-        RigEclipseCaseData* eclipseCase = rimEclipseCase->eclipseCaseData();
-        if ( !eclipseCase ) return;
-
-        RigCaseCellResultsData* cellResultsData = eclipseCase->results( this->porosityModel() );
-
-        size_t maxTimeStepCount = cellResultsData->maxTimeStepCount();
-        if ( this->isTernarySaturationSelected() && maxTimeStepCount > 1 )
-        {
-            RigCaseCellResultsData* gridCellResults = this->currentGridCellResults();
-            {
-                RigEclipseResultAddress resAddr( RiaDefines::ResultCatType::DYNAMIC_NATIVE, RiaResultNames::soil() );
-
-                if ( gridCellResults->ensureKnownResultLoaded( resAddr ) )
-                {
-                    double globalMin = 0.0;
-                    double globalMax = 1.0;
-                    double localMin  = 0.0;
-                    double localMax  = 1.0;
-
-                    cellResultsData->minMaxCellScalarValues( resAddr, globalMin, globalMax );
-                    cellResultsData->minMaxCellScalarValues( resAddr, currentTimeStep, localMin, localMax );
-
-                    ternaryLegendConfigToUpdate->setAutomaticRanges( RimTernaryLegendConfig::TERNARY_SOIL_IDX,
-                                                                     globalMin,
-                                                                     globalMax,
-                                                                     localMin,
-                                                                     localMax );
-                }
-            }
-
-            {
-                RigEclipseResultAddress resAddr( RiaDefines::ResultCatType::DYNAMIC_NATIVE, RiaResultNames::sgas() );
-
-                if ( gridCellResults->ensureKnownResultLoaded( resAddr ) )
-                {
-                    double globalMin = 0.0;
-                    double globalMax = 1.0;
-                    double localMin  = 0.0;
-                    double localMax  = 1.0;
-
-                    cellResultsData->minMaxCellScalarValues( resAddr, globalMin, globalMax );
-                    cellResultsData->minMaxCellScalarValues( resAddr, currentTimeStep, localMin, localMax );
-
-                    ternaryLegendConfigToUpdate->setAutomaticRanges( RimTernaryLegendConfig::TERNARY_SGAS_IDX,
-                                                                     globalMin,
-                                                                     globalMax,
-                                                                     localMin,
-                                                                     localMax );
-                }
-            }
-
-            {
-                RigEclipseResultAddress resAddr( RiaDefines::ResultCatType::DYNAMIC_NATIVE, RiaResultNames::swat() );
-
-                if ( gridCellResults->ensureKnownResultLoaded( resAddr ) )
-                {
-                    double globalMin = 0.0;
-                    double globalMax = 1.0;
-                    double localMin  = 0.0;
-                    double localMax  = 1.0;
-
-                    cellResultsData->minMaxCellScalarValues( resAddr, globalMin, globalMax );
-                    cellResultsData->minMaxCellScalarValues( resAddr, currentTimeStep, localMin, localMax );
-
-                    ternaryLegendConfigToUpdate->setAutomaticRanges( RimTernaryLegendConfig::TERNARY_SWAT_IDX,
-                                                                     globalMin,
-                                                                     globalMax,
-                                                                     localMin,
-                                                                     localMax );
-                }
-            }
-        }
+        RimEclipseResultDefinitionTools::updateTernaryLegend( currentGridCellResults(), ternaryLegendConfigToUpdate, currentTimeStep );
     }
 }
 
@@ -2256,6 +1787,14 @@ void RimEclipseResultDefinition::updateLegendTitle( RimRegularLegendConfig* lege
     }
 
     legendConfig->setTitle( title );
+}
+
+//--------------------------------------------------------------------------------------------------
+///
+//--------------------------------------------------------------------------------------------------
+bool RimEclipseResultDefinition::showOnlyVisibleCategoriesInLegend() const
+{
+    return m_showOnlyVisibleCategoriesInLegend();
 }
 
 //--------------------------------------------------------------------------------------------------
