+++ ApplicationLibCode/ProjectDataModel/RimEclipseResultDefinitionTools.h
@@ -26,25 +26,37 @@
 
 using RD = RimEclipseResultDefinition;
 
-class RimEclipseResultDefinitionTools
+namespace RimEclipseResultDefinitionTools
 {
-public:
-    static bool                         isDivideByCellFaceAreaPossible( const QString& resultName );
-    static RD::FlowTracerSelectionState getFlowTracerSelectionState( bool                        isInjector,
-                                                                     RD::FlowTracerSelectionType tracerSelectionType,
-                                                                     RimFlowDiagSolution*        flowDiagSolution,
-                                                                     size_t                      selectedTracerCount );
-    static QStringList getResultNamesForResultType( RiaDefines::ResultCatType resultCatType, const RigCaseCellResultsData* results );
+bool isDivideByCellFaceAreaPossible( const QString& resultName );
 
-    static QString timeOfFlightString( RD::FlowTracerSelectionState injectorState, RD::FlowTracerSelectionState producerState, bool shorter );
-    static QString
-        maxFractionTracerString( RD::FlowTracerSelectionState injectorState, RD::FlowTracerSelectionState producerState, bool shorter );
+RD::FlowTracerSelectionState getFlowTracerSelectionState( bool                        isInjector,
+                                                          RD::FlowTracerSelectionType tracerSelectionType,
+                                                          RimFlowDiagSolution* const  flowDiagSolution,
+                                                          size_t                      selectedTracerCount );
 
-    static QString selectedTracersString( RD::FlowTracerSelectionState injectorState,
-                                          RD::FlowTracerSelectionState producerState,
-                                          const std::vector<QString>&  selectedInjectors,
-                                          const std::vector<QString>&  selectedProducers,
-                                          int                          maxTracerStringLength );
+QStringList getResultNamesForResultType( RiaDefines::ResultCatType resultCatType, RigCaseCellResultsData* const results );
 
-    static QString getInputPropertyFileName( const RimEclipseCase* eclipseCase, const QString& resultName );
-};
+QString timeOfFlightString( RD::FlowTracerSelectionState injectorState, RD::FlowTracerSelectionState producerState, bool shorter );
+QString maxFractionTracerString( RD::FlowTracerSelectionState injectorState, RD::FlowTracerSelectionState producerState, bool shorter );
+
+QString selectedTracersString( RD::FlowTracerSelectionState injectorState,
+                               RD::FlowTracerSelectionState producerState,
+                               const std::vector<QString>&  selectedInjectors,
+                               const std::vector<QString>&  selectedProducers,
+                               int                          maxTracerStringLength );
+
+QString getInputPropertyFileName( RimEclipseCase* const eclipseCase, const QString& resultName );
+
+void updateTernaryLegend( RigCaseCellResultsData* const cellResultsData, RimTernaryLegendConfig* const ternaryLegendConfigToUpdate, int timeStep );
+
+void updateLegendForFlowDiagnostics( RD* const resultDefinition, RimRegularLegendConfig* const legendConfigToUpdate, int timeStep );
+
+void updateCellResultLegend( RD* const resultDefinition, RimRegularLegendConfig* const legendConfigToUpdate, int timeStep );
+
+QList<caf::PdmOptionItemInfo> calcOptionsForVariableUiFieldStandard( RiaDefines::ResultCatType     resultCatType,
+                                                                     RigCaseCellResultsData* const results,
+                                                                     bool                          showDerivedResultsFirst,
+                                                                     bool                          addPerCellFaceOptionItems,
+                                                                     bool                          enableTernary );
+}; // namespace RimEclipseResultDefinitionTools
